Revert "chore: remove global loading spinner"
