+++ studio/components/interfaces/App/GoTrueWrapper/GoTrueWrapper.utils.ts
@@ -1,0 +1,7 @@
+export function doesTokenDataExist() {
+  // ignore if server-side
+  if (typeof window === 'undefined') return false
+  // check tokenData on localstorage
+  const tokenData = window?.localStorage['supabase.auth.token']
+  return tokenData != undefined && typeof tokenData === 'string'
+}
