+++ studio/components/interfaces/App/index.ts
@@ -1,5 +1,6 @@
 import PortalToast from './PortalToast'
+import GoTrueWrapper from './GoTrueWrapper/GoTrueWrapper'
 import RouteValidationWrapper from './RouteValidationWrapper'
 import AppBannerWrapper from './AppBannerWrapper'
 
-export { PortalToast, RouteValidationWrapper, AppBannerWrapper }
+export { PortalToast, GoTrueWrapper, RouteValidationWrapper, AppBannerWrapper }
