+++ studio/components/interfaces/Auth/AuthProvidersForm/AuthProvidersForm.types.ts
@@ -12,15 +12,14 @@ export interface Provider {
     [x: string]: {
       title: string
       type: 'boolean' | 'string' | 'select' | 'number'
+      description?: string
+      descriptionOptional?: string
+      units?: string
       enum: Enum[]
       show: {
         key: string
         matches: string
       }
-      description?: string
-      descriptionOptional?: string
-      units?: string
-      isSecret?: boolean
     }
   }
   validationSchema: any // todo: use Yup type
