+++ studio/components/interfaces/Functions/CommandRender.tsx
@@ -4,10 +4,10 @@ import { useState } from 'react'
 const CommandRender = ({ commands }: any) => {
   return (
     <div className="space-y-4">
-      {commands.map((item: any, idx: number) => {
+      {commands.map((item: any) => {
         const [isCopied, setIsCopied] = useState(false)
         return (
-          <div key={`command-${idx}`} className="space-y-1">
+          <div className="space-y-1">
             <span className="font-mono text-sm text-scale-900">{`> ${item.comment}`}</span>
             <div className="flex items-center gap-2">
               <div className="flex gap-2 text-scale-1200 font-mono text-sm font-normal">
