+++ studio/components/interfaces/Organization/index.ts
@@ -1,13 +1,4 @@
 import BillingSettings from './BillingSettings/BillingSettings'
 import InvoicesSettings from './InvoicesSettings/InvoicesSettings'
-import InviteMemberModal from './InviteMemberModal'
-import DeleteOrganizationButton from './DeleteOrganizationButton'
-import OwnerDropdown from './OwnerDropdown'
 
-export {
-  BillingSettings,
-  InvoicesSettings,
-  InviteMemberModal,
-  DeleteOrganizationButton,
-  OwnerDropdown,
-}
+export { BillingSettings, InvoicesSettings }
