+++ studio/components/to-be-cleaned/Docs/Description.js
@@ -44,8 +44,8 @@ export default function Description({ content, metadata, onChange = () => {} }) 
       let query = ''
       let description = value.replaceAll("'", "''")
       if (table && column)
-        query = `comment on column public."${table}"."${column}" is '${description}';`
-      if (table && !column) query = `comment on table public."${table}" is '${description}';`
+        query = `comment on column public.${table}.${column} is '${description}';`
+      if (table && !column) query = `comment on table public.${table} is '${description}';`
       if (rpc) query = `comment on function ${rpc} is '${description}';`
 
       if (query) {
