+++ studio/lib/common/fetch/delete.ts
@@ -10,7 +10,7 @@ export async function delete_<T = any>(
   const requestId = uuidv4()
   try {
     const { headers: optionHeaders, ...otherOptions } = options ?? {}
-    const headers = await constructHeaders(requestId, optionHeaders)
+    const headers = constructHeaders(requestId, optionHeaders)
     const response = await fetch(url, {
       method: 'DELETE',
       body: JSON.stringify(data),
