+++ studio/lib/common/fetch/get.ts
@@ -9,7 +9,7 @@ export async function get<T = any>(
   const requestId = uuidv4()
   try {
     const { headers: optionHeaders, ...otherOptions } = options ?? {}
-    const headers = await constructHeaders(requestId, optionHeaders)
+    const headers = constructHeaders(requestId, optionHeaders)
     const response = await fetch(url, {
       method: 'GET',
       credentials: 'include',
@@ -34,7 +34,7 @@ export async function getWithTimeout<T = any>(
     const controller = new AbortController()
     const id = setTimeout(() => controller.abort(), timeout)
     const { headers: optionHeaders, ...otherOptions } = options ?? {}
-    const headers = await constructHeaders(requestId, optionHeaders)
+    const headers = constructHeaders(requestId, optionHeaders)
     const response = await fetch(url, {
       method: 'GET',
       credentials: 'include',
