+++ studio/lib/common/fetch/head.ts
@@ -10,7 +10,7 @@ export async function head<T = any>(
   const requestId = uuidv4()
   try {
     const { headers: optionHeaders, ...otherOptions } = options ?? {}
-    const headers = await constructHeaders(requestId, optionHeaders)
+    const headers = constructHeaders(requestId, optionHeaders)
     const response = await fetch(url, {
       method: 'HEAD',
       credentials: 'include',
@@ -36,7 +36,7 @@ export async function headWithTimeout<T = any>(
     const controller = new AbortController()
     const id = setTimeout(() => controller.abort(), timeout)
     const { headers: optionHeaders, ...otherOptions } = options ?? {}
-    const headers = await constructHeaders(requestId, optionHeaders)
+    const headers = constructHeaders(requestId, optionHeaders)
     const response = await fetch(url, {
       method: 'HEAD',
       credentials: 'include',
