+++ studio/package.json
@@ -27,7 +27,7 @@
     "@supabase/postgres-meta": "^0.26.1",
     "@supabase/react-data-grid": "^7.1.0-beta.7",
     "@supabase/shared-types": "^0.1.24",
-    "@supabase/supabase-js": "^1.36.0-next.9",
+    "@supabase/supabase-js": "^1.35.3",
     "@supabase/ui": "0.37.0-alpha.71",
     "ajv": "^8.6.3",
     "analytics-node": "^3.5.0",
