+++ studio/pages/org/[slug]/settings.tsx
@@ -1,46 +1,55 @@
-import Image from 'next/image'
 import { createContext, useEffect, useContext, useState } from 'react'
 import { useRouter } from 'next/router'
 import { observer, useLocalObservable } from 'mobx-react-lite'
 import { toJS } from 'mobx'
-import { pluckJsonSchemaFields, pluckObjectFields } from 'lib/helpers'
+import { pluckJsonSchemaFields, pluckObjectFields, timeout } from 'lib/helpers'
 import { AutoField } from 'uniforms-bootstrap4'
 import { organizations } from 'stores/jsonSchema'
 import {
   Loading,
   Button,
   Badge,
+  IconMoreHorizontal,
   Tabs,
   Typography,
+  IconTrash,
   Alert,
   Input,
+  Dropdown,
+  Modal,
   IconSearch,
   IconUser,
+  Form,
 } from '@supabase/ui'
 
-import { useOrganizationDetail, useStore, withAuth } from 'hooks'
-import { NextPageWithLayout, Project } from 'types'
 import { API_URL } from 'lib/constants'
-import { post, patch } from 'lib/common/fetch'
-
-import Panel from 'components/ui/Panel'
+import { useOrganizationDetail, useStore, withAuth } from 'hooks'
+import { post, delete_, patch } from 'lib/common/fetch'
 import { AccountLayoutWithoutAuth } from 'components/layouts'
-import {
-  BillingSettings,
-  InvoicesSettings,
-  OwnerDropdown,
-  InviteMemberModal,
-  DeleteOrganizationButton,
-} from 'components/interfaces/Organization'
+import { BillingSettings, InvoicesSettings } from 'components/interfaces/Organization'
 
 import Table from 'components/to-be-cleaned/Table'
-import SchemaFormPanel from 'components/to-be-cleaned/forms/SchemaFormPanel'
+import Panel from 'components/ui/Panel'
 import { confirmAlert } from 'components/to-be-cleaned/ModalsDeprecated/ConfirmModal'
-import { isInviteExpired } from 'components/interfaces/Organization/Organization.utils'
+import InviteMemberModal from 'components/interfaces/Organization/InviteMemberModal'
+import TextConfirmModal from 'components/ui/Modals/TextConfirmModal'
+import SchemaFormPanel from 'components/to-be-cleaned/forms/SchemaFormPanel'
+import { Member, NextPageWithLayout, Project } from 'types'
+import Image from 'next/image'
 
 // [Joshen] Low prio refactor: Bring out general and team settings into their own components too
 
-export const PageContext = createContext(null)
+const PageContext = createContext(null)
+
+// Invite is expired if older than 24hrs
+function inviteExpired(timestamp: Date) {
+  const inviteDate = new Date(timestamp)
+  const now = new Date()
+  var timeBetween = now.valueOf() - inviteDate.valueOf()
+  if (timeBetween / 1000 / 60 / 60 < 24) {
+    return true
+  }
+}
 
 const OrgSettingsLayout = withAuth(
   observer(({ children }) => {
@@ -170,10 +179,22 @@ const TabsView = observer(() => {
         </section>
         <nav className="">
           <Tabs onChange={(id: any) => setSelectedTab(id)} type="underlined">
-            <Tabs.Panel id="GENERAL" label="General" />
-            <Tabs.Panel id="TEAM" label="Team" />
-            <Tabs.Panel id="BILLING" label="Billing" />
-            <Tabs.Panel id="INVOICES" label="Invoices" />
+            {/* @ts-ignore */}
+            <Tabs.Panel id="GENERAL" label="General">
+              <></>
+            </Tabs.Panel>
+            {/* @ts-ignore */}
+            <Tabs.Panel id="TEAM" label="Team">
+              <></>
+            </Tabs.Panel>
+            {/* @ts-ignore */}
+            <Tabs.Panel id="BILLING" label="Billing">
+              <></>
+            </Tabs.Panel>
+            {/* @ts-ignore */}
+            <Tabs.Panel id="INVOICES" label="Invoices">
+              <></>
+            </Tabs.Panel>
           </Tabs>
         </nav>
       </div>
@@ -268,10 +289,120 @@ const OrgDeletePanel = observer(() => {
           <p className="text-red-900">
             Make sure you have made a backup if you want to keep your data
           </p>
-          <DeleteOrganizationButton />
+          <OrgDeleteModal />
         </Alert>
       </Panel.Content>
     </Panel>
+  )
+})
+
+const OrgDeleteModal = observer(() => {
+  const PageState: any = useContext(PageContext)
+  const router = useRouter()
+  const { ui } = useStore()
+
+  const { slug: orgSlug, name: orgName } = PageState.organization
+
+  const [isOpen, setIsOpen] = useState(false)
+  const [value, setValue] = useState('')
+
+  function toggle() {
+    setIsOpen(!isOpen)
+  }
+
+  return (
+    <>
+      <div className="mt-2">
+        <Button onClick={toggle} type="danger">
+          Delete organization
+        </Button>
+      </div>
+      <Modal
+        visible={isOpen}
+        onCancel={toggle}
+        header={
+          <div className="flex items-baseline gap-2">
+            <h5 className="text-scale-1200 text-sm">Delete organisation</h5>
+            <span className="text-scale-900 text-xs">Are you sure?</span>
+          </div>
+        }
+        size="small"
+        hideFooter
+        closable
+      >
+        <Form
+          initialValues={{
+            orgName: '',
+          }}
+          validateOnBlur
+          onSubmit={async (values: any, { setSubmitting }: any) => {
+            setSubmitting(true)
+            const response = await delete_(`${API_URL}/organizations/${orgSlug}`)
+            if (response.error) {
+              ui.setNotification({
+                category: 'error',
+                message: `Failed to delete organization: ${response.error.message}`,
+              })
+              setSubmitting(false)
+            } else {
+              PageState.onOrgDeleted(PageState.organization)
+              setSubmitting(false)
+              router.push('/')
+            }
+          }}
+          validate={(values) => {
+            const errors: any = {}
+            if (!values.orgName) {
+              errors.orgName = 'Enter the name of the organization.'
+            }
+            if (values.orgName !== orgSlug) {
+              errors.orgName = 'Value entered does not match name of the organization.'
+            }
+            return errors
+          }}
+        >
+          {({ isSubmitting }: { isSubmitting: boolean }) => (
+            <div className="space-y-4 py-3">
+              <Modal.Content>
+                <p className="text-scale-900 text-sm">
+                  This action <span className="text-scale-1200">cannot</span> be undone. This will
+                  permanently delete the <span className="text-scale-1200">{orgName}</span>{' '}
+                  organization and remove all of its projects.
+                </p>
+              </Modal.Content>
+              <Modal.Seperator />
+              <Modal.Content>
+                <Input
+                  id="orgName"
+                  label={
+                    <span>
+                      Please type <Typography.Text strong>{orgSlug}</Typography.Text> to confirm
+                    </span>
+                  }
+                  onChange={(e) => setValue(e.target.value)}
+                  value={value}
+                  placeholder="Type in the orgnaization name"
+                  className="w-full"
+                />
+              </Modal.Content>
+              <Modal.Seperator />
+              <Modal.Content>
+                <Button
+                  type="danger"
+                  htmlType="submit"
+                  loading={isSubmitting}
+                  size="small"
+                  block
+                  danger
+                >
+                  I understand, delete this organization
+                </Button>
+              </Modal.Content>
+            </div>
+          )}
+        </Form>
+      </Modal>
+    </>
   )
 })
 
@@ -401,8 +532,8 @@ const MembersView = observer(() => {
 
                 <Table.td>
                   {x.invited_id && (
-                    <Badge color={isInviteExpired(x.invited_at) ? 'yellow' : 'red'}>
-                      {isInviteExpired(x.invited_at) ? 'Invited' : 'Expired'}
+                    <Badge color={inviteExpired(x.invited_at) ? 'yellow' : 'red'}>
+                      {inviteExpired(x.invited_at) ? 'Invited' : 'Expired'}
                     </Badge>
                   )}
                 </Table.td>
@@ -438,6 +569,194 @@ const MembersView = observer(() => {
           ]}
         />
       </Loading>
+    </div>
+  )
+})
+
+const OwnerDropdown = observer(({ members, member }: any) => {
+  const PageState: any = useContext(PageContext)
+  const { ui } = useStore()
+  const { mutateOrgMembers } = useOrganizationDetail(ui.selectedOrganization?.slug || '')
+  const [isOpen, setIsOpen] = useState(false)
+  const [loading, setLoading] = useState(false)
+
+  // handle modal visibility
+  const [ownerTransferIsVisble, setOwnerTransferIsVisble] = useState(false)
+
+  const { id: orgId, slug: orgSlug, stripe_customer_id, name: orgName } = PageState.organization
+
+  async function handleMemberDelete() {
+    setIsOpen(false)
+    await timeout(200)
+
+    confirmAlert({
+      title: 'Confirm to remove',
+      message: `This is permanent! Are you sure you want to remove ${member.profile.primary_email}?`,
+      onAsyncConfirm: async () => {
+        setLoading(true)
+        const response = await delete_(`${API_URL}/organizations/${orgSlug}/members/remove`, {
+          member_id: member.id,
+        })
+        if (response.error) {
+          ui.setNotification({
+            category: 'error',
+            message: `Failed to delete user: ${response.error.message}`,
+          })
+          setLoading(false)
+        } else {
+          const updatedMembers = members.filter((x: any) => x.id !== member.id)
+          mutateOrgMembers(updatedMembers)
+          ui.setNotification({ category: 'success', message: 'Successfully removed member' })
+        }
+      },
+    })
+  }
+
+  async function handleTransfer() {
+    setLoading(true)
+
+    const response = await post(`${API_URL}/organizations/${orgSlug}/transfer`, {
+      org_id: orgId,
+      member_id: member.id,
+    })
+    if (response.error) {
+      ui.setNotification({
+        category: 'error',
+        message: `Failed to transfer ownership: ${response.error.message}`,
+      })
+      setLoading(false)
+    } else {
+      const updatedMembers = [...members]
+      const oldOwner = updatedMembers.find((x) => x.is_owner == true)
+      if (oldOwner) oldOwner.is_owner = false
+      const newOwner = updatedMembers.find((x) => x.id == member.id)
+      if (newOwner) newOwner.is_owner = true
+      mutateOrgMembers(updatedMembers)
+      setOwnerTransferIsVisble(false)
+      ui.setNotification({ category: 'success', message: 'Successfully transfered organization' })
+    }
+  }
+
+  async function handleResendInvite(member: Member) {
+    setLoading(true)
+
+    const response = await post(`${API_URL}/organizations/${orgSlug}/members/invite`, {
+      invited_email: member.profile.primary_email,
+      owner_id: member.invited_id,
+    })
+
+    if (response.error) {
+      ui.setNotification({
+        category: 'error',
+        message: `Failed to resend invitation: ${response.error.message}`,
+      })
+      setLoading(false)
+    } else {
+      const updatedMembers = [...members]
+      mutateOrgMembers(updatedMembers)
+      ui.setNotification({ category: 'success', message: 'Resent the invitation.' })
+      setLoading(false)
+    }
+  }
+
+  async function handleRevokeInvitation(id: number) {
+    setLoading(true)
+
+    const response = await delete_(
+      `${API_URL}/organizations/${orgSlug}/members/invite?invited_id=${id}`,
+      {}
+    )
+
+    if (response.error) {
+      ui.setNotification({
+        category: 'error',
+        message: `Failed to revoke invitation: ${response.error.message}`,
+      })
+      setLoading(false)
+    } else {
+      const updatedMembers = [...members]
+      mutateOrgMembers(updatedMembers)
+      ui.setNotification({ category: 'success', message: 'Successfully revoked the invitation.' })
+    }
+  }
+
+  return (
+    <div className="flex items-center justify-end">
+      <Dropdown
+        side="bottom"
+        align="end"
+        overlay={
+          <>
+            {!member.invited_at && (
+              <Dropdown.Item onClick={() => setOwnerTransferIsVisble(!ownerTransferIsVisble)}>
+                <div className="flex flex-col">
+                  <p>Make owner</p>
+                  <p className="block opacity-50">Transfer ownership of "{orgName}"</p>
+                </div>
+              </Dropdown.Item>
+            )}
+
+            {member.invited_at && (
+              <>
+                <Dropdown.Item onClick={() => handleRevokeInvitation(member.invited_id)}>
+                  <div className="flex flex-col">
+                    <p>Cancel invitation</p>
+                    <p className="block opacity-50">Revoke this invitation.</p>
+                  </div>
+                </Dropdown.Item>
+
+                {!inviteExpired(member.invited_at) && (
+                  <>
+                    <Dropdown.Seperator />
+                    <Dropdown.Item onClick={() => handleResendInvite(member)}>
+                      <div className="flex flex-col">
+                        <p>Resend invitation</p>
+                        <p className="block opacity-50">Invites expire after 24hrs.</p>
+                      </div>
+                    </Dropdown.Item>
+                  </>
+                )}
+              </>
+            )}
+
+            {!member.invited_at && (
+              <>
+                <Dropdown.Seperator />
+                <Dropdown.Item icon={<IconTrash size="tiny" />} onClick={handleMemberDelete}>
+                  Remove member
+                </Dropdown.Item>
+              </>
+            )}
+          </>
+        }
+      >
+        <Button
+          as="span"
+          disabled={loading}
+          loading={loading}
+          type="text"
+          icon={<IconMoreHorizontal />}
+        ></Button>
+      </Dropdown>
+
+      <TextConfirmModal
+        title="Transfer organization"
+        visible={ownerTransferIsVisble}
+        confirmString={orgSlug}
+        loading={loading}
+        confirmLabel="I understand, transfer ownership"
+        confirmPlaceholder="Type in name of orgnization"
+        onCancel={() => setOwnerTransferIsVisble(!ownerTransferIsVisble)}
+        onConfirm={handleTransfer}
+        alert="Payment methods such as credit cards will also be transferred. You may want to delete credit card information first before transferring."
+        text={
+          <span>
+            By transferring this organization, it will be solely owned by{' '}
+            <span className="font-medium dark:text-white">{member.profile?.username}</span>, they
+            will also be able to remove you from the organization as a member
+          </span>
+        }
+      />
     </div>
   )
 })
