+++ studio/pages/project/[ref]/editor/[id].tsx
@@ -36,9 +36,7 @@ const TableEditorPage: NextPage = () => {
   const selectedTable = !isNaN(Number(id))
     ? // @ts-ignore
       tables.find((table) => table.id === Number(id))
-    : id !== undefined
-    ? tryParseJson(Base64.decode(id))
-    : undefined
+    : tryParseJson(Base64.decode(id))
 
   useEffect(() => {
     if (selectedTable && 'schema' in selectedTable) {
