+++ studio/stores/authConfig/schema/AuthProviders/AuthProvidersFormValidation.tsx
@@ -86,7 +86,6 @@ const PROVIDER_PHONE = {
     SMS_TWILIO_AUTH_TOKEN: {
       type: 'string',
       title: 'Twilio Auth Token',
-      isSecret: true,
       show: {
         key: 'SMS_PROVIDER',
         matches: 'twilio',
@@ -312,7 +311,6 @@ Client identifier when authenticating or validating users.
 The secret key is a JWT token that must be generated. 
 [Learn more](https://supabase.com/docs/guides/auth/auth-apple#generate-a-client_secret)`,
       type: 'string',
-      isSecret: true,
     },
   },
   validationSchema: object().shape({
@@ -358,7 +356,6 @@ const EXTERNAL_PROVIDER_AZURE = {
       // [TODO] Change docs
       title: 'Secret ID',
       type: 'string',
-      isSecret: true,
     },
     EXTERNAL_AZURE_URL: {
       // [TODO] Change docs
@@ -403,7 +400,6 @@ const EXTERNAL_PROVIDER_BITBUCKET = {
     EXTERNAL_BITBUCKET_SECRET: {
       title: 'Secret',
       type: 'string',
-      isSecret: true,
     },
   },
   validationSchema: object().shape({
@@ -441,7 +437,6 @@ const EXTERNAL_PROVIDER_DISCORD = {
     EXTERNAL_DISCORD_SECRET: {
       title: 'Client Secret',
       type: 'string',
-      isSecret: true,
     },
   },
   validationSchema: object().shape({
@@ -479,7 +474,6 @@ const EXTERNAL_PROVIDER_FACEBOOK = {
     EXTERNAL_FACEBOOK_SECRET: {
       title: 'Facebook secret',
       type: 'string',
-      isSecret: true,
     },
   },
   validationSchema: object().shape({
@@ -517,7 +511,6 @@ const EXTERNAL_PROVIDER_GITHUB = {
     EXTERNAL_GITHUB_SECRET: {
       title: 'Client Secret',
       type: 'string',
-      isSecret: true,
     },
   },
   validationSchema: object().shape({
@@ -557,7 +550,6 @@ const EXTERNAL_PROVIDER_GITLAB = {
     EXTERNAL_GITLAB_SECRET: {
       title: 'Secret',
       type: 'string',
-      isSecret: true,
     },
     EXTERNAL_GITLAB_URL: {
       title: 'Self Hosted GitLab URL',
@@ -603,7 +595,6 @@ const EXTERNAL_PROVIDER_GOOGLE = {
     EXTERNAL_GOOGLE_SECRET: {
       title: 'Client Secret',
       type: 'string',
-      isSecret: true,
     },
   },
   validationSchema: object().shape({
@@ -642,7 +633,6 @@ const EXTERNAL_PROVIDER_KEYCLOAK = {
     EXTERNAL_KEYCLOAK_SECRET: {
       title: 'Secret',
       type: 'string',
-      isSecret: true,
     },
     EXTERNAL_KEYCLOAK_URL: {
       title: 'Realm URL',
@@ -693,7 +683,6 @@ const EXTERNAL_PROVIDER_LINKEDIN = {
     EXTERNAL_LINKEDIN_SECRET: {
       title: 'API Secret Key',
       type: 'string',
-      isSecret: true,
     },
   },
   validationSchema: object().shape({
@@ -731,7 +720,6 @@ const EXTERNAL_PROVIDER_NOTION = {
     EXTERNAL_NOTION_SECRET: {
       title: 'OAuth client secret',
       type: 'string',
-      isSecret: true,
     },
   },
   validationSchema: object().shape({
@@ -769,7 +757,6 @@ const EXTERNAL_PROVIDER_TWITCH = {
     EXTERNAL_TWITCH_SECRET: {
       title: 'Client secret',
       type: 'string',
-      isSecret: true,
     },
   },
   validationSchema: object().shape({
@@ -807,7 +794,6 @@ const EXTERNAL_PROVIDER_TWITTER = {
     EXTERNAL_TWITTER_SECRET: {
       title: 'API Secret Key',
       type: 'string',
-      isSecret: true,
     },
   },
   validationSchema: object().shape({
@@ -845,7 +831,6 @@ const EXTERNAL_PROVIDER_SLACK = {
     EXTERNAL_SLACK_SECRET: {
       title: 'Client Secret',
       type: 'string',
-      isSecret: true,
     },
   },
   validationSchema: object().shape({
@@ -883,7 +868,6 @@ const EXTERNAL_PROVIDER_SPOTIFY = {
     EXTERNAL_SPOTIFY_SECRET: {
       title: 'Client Secret',
       type: 'string',
-      isSecret: true,
     },
   },
   validationSchema: object().shape({
@@ -925,7 +909,6 @@ const EXTERNAL_PROVIDER_WORKOS = {
     EXTERNAL_WORKOS_SECRET: {
       title: 'Secret Key',
       type: 'string',
-      isSecret: true,
     },
   },
   validationSchema: object().shape({
@@ -970,7 +953,6 @@ const EXTERNAL_PROVIDER_ZOOM = {
     EXTERNAL_ZOOM_SECRET: {
       title: 'Client secret',
       type: 'string',
-      isSecret: true,
     },
   },
   validationSchema: object().shape({
